+++ tsclient/src/OtherStuff/index.ts
@@ -1,5 +1,5 @@
-import holding from "./Holding"
+import holding from "./Init"
 import * as Data from "./Data"
-import something from "./Something"
+import something from "./Order"
 
 export {holding, Data,something}
